requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.3.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
dulwich==0.21.7
eval_type_backport==0.2.0
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
fastjsonschema==2.21.1
filelock==3.15.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.9.0
h11==0.14.0
http-exceptions==0.2.10
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.23.4
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
llama-cloud==0.1.5
llama-index==0.11.23
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.23
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.3.0
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.14
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy
openai==1.58.1
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.27.2
ptyprocess==0.7.0
pyarrow==16.1.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pypdf==5.1.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.41.1
striprtf==0.0.26
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.8.0
together==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.13.2
toolhouse==1.2.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.67.1
trove-classifiers==2024.10.21.16
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.28.0
watchdog==4.0.1
wrapt==1.16.0
xattr==1.1.0
yarl==1.9.4
pyperclip